+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 
 ### Changed
 - [Document] Introduce buffering to accumulate markdown chunks before parsing
+- Fixed pipeline execution when `:markdown` is passed in `MDEx.new/1`
+
+### Deprecated
+- [Document] Deprecate `MDEx.Document.parse_markdown/2` and `MDEx.Document.parse_markdown!/2` in favor of `MDEx.parse_document/2` or `MDEx.Document.put_markdown/1`
 
 ## 0.9.1 - 2025-09-30